Fix batch edit to properly handle non-web pic formats
Previously, batch-update-pic-info! was infering the pic path from
thumbnail path.  This doesn't work for non-web formats, since the
thumbnail _is_ in a web format, so it is not possible to determine the
original pic path (the original extension is lost once the thumbnail
is created with a web-compatible extension).

To fix this, we add a data-pic-path attribute to the <img> tag for
thumbnails.  data-pic-path contains the original pic path.