batch edit: add option to specify whether data must be overwritten or not
Instead of blindly overwriting pictures' data, let the user specify if
he/she wants to overwrite data or not.

When overwrite is selected, the submitted data will completely
overwrite all the selected pictures' fields. Existing associations to
albuns/tags will be replaced by the submitted ones.  If overwrite is
selected, empty template fields will remove the corresponding values
of the selected pictures.

When overwrite is not selected (default), only fields that are empty
will be filled with the template data.  Albums and tags will be added
to the existing ones bound to the selected pictures.  If overwrite is
not selected, submitting empty template data doesn't remove pictures'
data.  Empty template data is just ignored.