Pagination: paginate items in memory
I guess paginating items in memory turns out to be better than
counting the total number of items in a separate query (the total
number of items is required by the pager).

The procedures from the DB API still accept `pagenum' as an optional
parameter, in case the current approach doesn't work well.